pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
description = "TradingView Signal Driven Trading System"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "greenlet>=3.0.0",  # SQLAlchemy async 运行依赖，缺则 engine.begin() 等报错
    "asyncpg>=0.29.0",  # PostgreSQL 异步驱动（项目使用 create_async_engine）
    "aiosqlite>=0.19.0",  # SQLite（测试环境推荐）
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "ccxt>=4.0.0",  # 交易所 API
    "apscheduler>=3.10.0",  # 进程内调度（禁止 Celery）
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",  # PR15a：RealOkxHttpClient（可注入 transport 做离线测试）
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",  # starlette.testclient / TestClient 依赖
    "black>=23.11.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.uvicorn]
workers = 1  # 单实例约束

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
